feat(core): Add firebase modules

+++ core/src/commonMain/kotlin/com/sjaindl/s11/core/firestore/player/model/Player.kt
@@ -9,5 +9,5 @@ data class Player(
     val position: Position,
     val imageRef: String?,
     val downloadUrl: String?,
-    val points: Map<String, Int>,
+    val points: Map<String, Float>,
 )
